fix(UI): Fix toggle button for Collision-avoidant Navigation
Before the fix, you needed another "toggle" of the button to see it move (in the backend, the setting is done from the first toggle, the real one) and so get the confirmation that the setting is applied.
Another approach was to toggle it then refresh / reload the page.
TODO: Investigate further why we need this workaround as other toggle switches work accordingly. Maybe it's because this one takes a little bit longer to be configured (combined bit values in a byte) on the robot.